project_1/requirements.txt
jupyter
matplotlib
numpy
scipy
seaborn
pillow
pytest
scikit-learn
